+++ src/SME.SGP.WebClient/src/paginas/Dashboard/DashboardDevolutivas/dashboardDevolutivasFiltros.js
@@ -223,6 +223,35 @@ const DashboardDevolutivasFiltros = () => {
     ServicoDashboardDevolutivas.atualizarFiltros('modalidade', valor);
   };
 
+  const obterUltimaConsolidacao = useCallback(async () => {
+    const resposta = await ServicoDashboardDevolutivas.obterUltimaConsolidacao(
+      anoLetivo
+    ).catch(e => erros(e));
+
+    if (resposta?.data) {
+      ServicoDashboardDevolutivas.atualizarFiltros(
+        'dataUltimaConsolidacao',
+        resposta.data
+      );
+    } else {
+      ServicoDashboardDevolutivas.atualizarFiltros(
+        'dataUltimaConsolidacao',
+        undefined
+      );
+    }
+  }, [anoLetivo]);
+
+  useEffect(() => {
+    if (anoLetivo) {
+      obterUltimaConsolidacao();
+    } else {
+      ServicoDashboardDevolutivas.atualizarFiltros(
+        'dataUltimaConsolidacao',
+        undefined
+      );
+    }
+  }, [anoLetivo, obterUltimaConsolidacao]);
+
   return (
     <>
       <div className="row">
